Update the API spec and overview to reflect ServingState intent. (#657)
* Update the API spec and overview to reflect ServingState intent.

I've reworded the description of a Revision in the overview docs to capture some of the key lifecycle stages a revision may be in.  I also found the existing wording confusing, since it talked about "history", which is more a property of configuration that is enabled by retirement, and I think the new wording better captures this.

I've updated the specification to include `servingState`, clarify where fields aren't available (pointing to issues), and tweak assorted cosmetic things (e.g. trailing whitespace).

Fixes: https://github.com/elafros/elafros/issues/644

* Clarify that blocking Revision creation is NYI (link to issue).

* Various code review updated from @evankanderson and @grantr.