Update vendored test-infra version
This repo has fallen behind others in the vendored version of
test-infra and as a result integration tests are erroring out.

I'm using `dep` version `v0.5.0` locally, which matches what other
Knative repositories are using for generation of their Gopkg.lock
files. That does cause a bit more noise to this PR as the `digest` and
`pruneopts` fields in Gopkg.lock get filled in.