Add hello world for Rust Nightly which uses Rocket
Rocket is nice because it makes the code smaller, and it makes
URL route handling very elegant and easy.

This example also includes a best practice for speeding up
Rust Docker builds, and shows how to do a multi-stage build
so you can deploy the final application within a Docker
scratch container, leading to a final production container
that's only a couple of MB or less.