Update metrics and ingress gateway documentation
This change updates the metrics, logging, and ingress documentation with
information from documents in the Knative Serving repository. This is an
effort to centralize our user documentation in a single place.

This should allow us to remove the following files from Knative Serving
as they now contain either duplicative or out-dated information.

* https://github.com/knative/serving/blob/master/docs/telemetry.md
*
https://github.com/knative/serving/blob/master/docs/setting-up-custom-ingress-gateway.md
*
https://github.com/knative/serving/blob/master/docs/setting-up-ingress-static-ip.md