Building and deploying the sample
replace this: To find the IP address for your service, use. If your cluster is new, it may take sometime for the service to get assigned an external IP address.

with this:
To find the IP address for your service, use.

Note If your cluster is new, it may take sometime for the service to get assigned an external IP address. Instead of rerunning the command, you can add --watch to the command below to view the component's status updates in real time. Use CTRL+C to exit watch mode.