Remove references to Istio yamls attached to releases.
In 0.6 we will stop shipping Istio yamls, as we are not in the business of
redistributing Istio (incl. supporting these yamls).  In preparation for that,
this change makes a few changes:
1. Update things to point to the 0.5.2 release
1. Point to our in-tree Istio files (which is what we were attaching)
1. For GKE, direct users to the Istio addon instead.
1. For GKE, add a pointer to CR-GKE (if folks wanted hosted).

This is a half-measure to avoid things breaking.  We need a page that points
folks to the Istio docs, and calls out any specialized instructions for
what Knative needs.