Cherry-pick batch (#1306)
* Fix typo Knatve -> Knative (#1196)

Signed-off-by: xichengliudui <1693291525@qq.com>

* docs: Remove `--save` from `npm install` (#1206)

See https://stackoverflow.com/a/19578808

* serving/samples/helloworld-php: render newline (#1224)

* serving/samples/helloworld-php: render newline

* Double quotes in README too

* Fix a couple broken links to removed samples. (#1242)

* Remove references to Istio yamls attached to releases. (#1243)

In 0.6 we will stop shipping Istio yamls, as we are not in the business of
redistributing Istio (incl. supporting these yamls).  In preparation for that,
this change makes a few changes:
1. Update things to point to the 0.5.2 release
1. Point to our in-tree Istio files (which is what we were attaching)
1. For GKE, direct users to the Istio addon instead.
1. For GKE, add a pointer to CR-GKE (if folks wanted hosted).

This is a half-measure to avoid things breaking.  We need a page that points
folks to the Istio docs, and calls out any specialized instructions for
what Knative needs.

* Fix command line to extract LATEST_HOSTNAME (#1258)

* address wrong manifest in gcppubsub sample (#1273)

* fix broken demo (#1277)