gRPC Go README command fix
This change removes an errant `cd` command which causes the rest of the commands to fail as they all assume you're at the root directory. Also adds a command to force the reader to open the file which needs changing, otherwise it's too easy to just bang in the command and have it fail.