Building and deploying the sample (#911)
* Building and deploying the sample

replace this: To find the IP address for your service, use. If your cluster is new, it may take sometime for the service to get assigned an external IP address.

with this:
To find the IP address for your service, use.

Note If your cluster is new, it may take sometime for the service to get assigned an external IP address. Instead of rerunning the command, you can add --watch to the command below to view the component's status updates in real time. Use CTRL+C to exit watch mode.

* Building and deploying the sample

Incorporating suggested changes. Moved Note below the command to find the IP address. Changed sentences to take the form of To do/make etc.
Created a Removing the sample section

* undo the added title: Removing the sample

Accidental additional section for ## Removing the sample app deployment
Add reference to External-IP in To make a request to your app

* Update serving/samples/helloworld-java/README.md

change sometime to some time

Co-Authored-By: janitha09 <janitha_j@hotmail.com>

* Remove $ prefix on commands

The knative style guide prefers this formatting.

* Clean up whitespace

* Revert line continuation, which is unnecessary