This adds a simple cloudevents-go sample. (#1987)
This example is intended to showcase how to start an event receiver as a Knative
Service, and then have that service either send a new event to a `$K_SINK` or
respond with it (e.g. for working with a Broker).

Once we're happy with this, we should broaden this to cover more languages.