Remove namespace definition from hello-world service.yaml
Remove the namespace definition from the service.yaml in each of the helloworld-*
samples.

Specifying the namespace in service.yaml expects there to exist a namespace with
the specified name, in this case 'default'.  When namespace is unspecified, the
service.yaml will default to using the namespace specified in the default $KUBECONFIG.