Create overview.md (#2160)
* Create overview.md

A document for the Concepts section that is purely a high level description of what the project is.  The main landing page for the site includes lots of market-y stuff and process stuff, and is likely to be confusing to someone who is really new to all this. Most of this is stolen, shamelessly, from knative.tips. It also points out that Build has been moved to its own project.  Since much of the documentation available on the wider web still indicates Build is part of Knative, it is valuable to explain where it went.

* Update docs/concepts/overview.md

Co-Authored-By: Sam O'Dell <31352624+samodell@users.noreply.github.com>

* Update docs/concepts/overview.md

Co-Authored-By: Sam O'Dell <31352624+samodell@users.noreply.github.com>

* Update docs/concepts/overview.md

Co-Authored-By: Sam O'Dell <31352624+samodell@users.noreply.github.com>

* Update docs/concepts/overview.md

Co-Authored-By: Sam O'Dell <31352624+samodell@users.noreply.github.com>

* Update docs/concepts/overview.md

Co-Authored-By: Sam O'Dell <31352624+samodell@users.noreply.github.com>

* Update docs/concepts/overview.md

Co-Authored-By: Sam O'Dell <31352624+samodell@users.noreply.github.com>

Co-authored-by: Sam O'Dell <31352624+samodell@users.noreply.github.com>