Drop the custom install guide.
With the addition of sources and channels to the "any" install guide, this should now be fully redundant.

The only places of which I'm aware that are NOT covered are being intentionally dropped due to missing samples (e.g. AWS SQS source), and have been handed off to their respective working group leads.