Update knative-api-specification-1.0.md
This document states : 
      “name	string       (Optional)	Must be one of "http1" or "h2c" (if supported). Defaults to "http1".

I was also following this document  Kantive spec for RevisionTemplateSpec here:
https://knative.dev/docs/reference/serving-api/#serving.knative.dev/v1.RevisionTemplateSpec
That documentation defers to k8s for the definition of  PodSpec Kubernetes core/v1.PodSpec. 
     documented here https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.18/#podspec-v1-core
  which refers to containers Container array 
     documented here  https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.18/#container-v1-core
  which refers to ports ContainerPort array
      documented here https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.18/#containerport-v1-core 
   which states:
name. string	If specified, this must be an IANA_SVC_NAME and unique within the pod. Each named port in a pod must have a unique name. Name for the port that can be referred to by services.

I think it is important to explicitly communicate that these 2 definition relate to each other (In my specific case I was working on https://github.com/knative/serving/issues/4283 and it is unclear where new values like "auto" can be added and whether keeping the IANA registered is important. )

I think it is important to express that  this name is a subset of values supported by k8s (must be an IANA_SVC_NAME) also currently the only 2 that may are supported by  Knative are  "http1" and "h2c"(optional)