Promote net-contour to stable.
All of the known remaining issues with Contour have been resolved, and I've been carefully monitoring testgrid and digging into any 503s (or other dataplane related failures).

As of me writing this:
1. The only failure in the net-contour testgrid is due to its workaround for https://github.com/knative-sandbox/net-certmanager/issues/44, which shouldn't practically manifest in the context of Serving.
2. The only failures in the contour leg on Serving testgrid are due to infrastructure (Go, etcd, GKE, resource exhaustion)

I will continue to track testgrid carefully as we head towards 0.18, but at present it is likely as solid as any of our ingress options.