helloworld shell sample: Switch to use CGI bin for the output
Also, httpd is now started by init for proper SIGTERM handling.
If you run the image itself within Docker, note that SIGINT (CTRL-C) is not captured.
Use `docker --init` for being able to stop the containe with CTRL-C
(This is not relevant when used for Knative)

Thanks to @csantanapr for the idea to switch to CGI.