helloworld shell sample: Switch to use CGI bin for the output (#3053)
* helloworld shell sample: Switch to use CGI bin for the output

Also, httpd is now started by init for proper SIGTERM handling.
If you run the image itself within Docker, note that SIGINT (CTRL-C) is not captured.
Use `docker --init` for being able to stop the containe with CTRL-C
(This is not relevant when used for Knative)

Thanks to @csantanapr for the idea to switch to CGI.

* chore: Removed dynamic parts from the script

It's back now to a simple echo, so that the CI tests pass that
compare on exactly the output of a request. Also this is probably
easier to understand for a user, too.

Based on review feedback.

* chore: Add missing file