Update Eventing README
Switch `service` to `resources` when referring to internal Knative Eventing.

Why: Disambiguate `service` to refer to an abstraction which has an http endpoint and primarily referring to a K8s Service or Knative Service.
Here, also calling Knative Eventing resources as `services` creates confusion even if the usage is technically correct.