Remove Gloo install instructions
As discussed in the recent TOC meeting, we want to be sure the Ingresses
we list in the install instructions are ones we're confident will work
well with Knative, to avoid folks having a bad first experience.
Currently we don't have good (any) CI coverage for Gloo, so removing it
from the list for the moment until we do.