update fork and knative/docs repo url
While going through the GitHub workflow for Knative documentation, The very first section is "Set up your local machine ", which contains the first point to create your fork of the knative/docs repo. So for the easy access, we can update the fork URL with the "fork" word and then can attach the knative/docs URL to the "knative/docs repo" which can be easy for the new contributors to easily access the repo.