Clarify pre-req versions for eventing-rabbitmq (#4518)
We had a query about which versions should be installed:
https://jira.eng.vmware.com/browse/PDY-649

Our aim is to support the latest released versions of the RabbitMQ
Cluster Operator & Messaging Topology Operator. I am highlighting the
Cert Manager option as that is the easiest way to get started, and is
also what we do in our acceptance testing.

Since this is a Getting Started, I have removed the deadLetterSink
option since it's optional. I don't think that optional properties have
a place in a Getting Started.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>