Update Node.js code samples w/ latest CloudEvents (#4865)
This commit updates two of the examples in this repository that use the
Node.js CloudEvents module. A newer version of the module has been released
and the examples have been updated to use it. The newer version allows us
to eliminate a dependency on axios, and simplify the example code.

Signed-off-by: Lance Ball <lball@redhat.com>