enhancement: add language packs documentation
This commit adds two files to the top level navigation for functions.
One covering the built-in templates, and the other covering language
packs. Note that it does _not_ include documentation around how to
create your own language packs. Rather, it links to the repository
documentation for the Language Pack Contract. This should eventually
also be added to the knative.dev site.

Fixes: https://github.com/knative/docs/issues/5287

Signed-off-by: Lance Ball <lball@redhat.com>