Fix e2e tutorial solution files and documentation.
Both function templates are updated to reflect the most recent version
of the function cli. The code was also updated to build and execute.

The `solution/setup.sh` was removed because it is empty.

Mentioned Kind/Minikube with a link to the local registry setup docs in
`solution/README.md`. This is needed for Camel K operator installation.

Add corrected Camel K operator installation to `slack-sink/README.md`
and to `solution.sh`. Also remove `kamel` cli installation check.

Fixed typo and whitespace in the Javascript code of the node-server.

Signed-off-by: Stanislav Jakuschevskij <sjakusch@redhat.com>