Fix section "3-Bad Word Service" in e2e tutorial (#6425)
* Fix section "3-Bad Word Service" in e2e tutorial

The latest func cli generates a different Python function template than
shown in the tutorial. There is no setup.py anymore and the project
uses a toml file for configuration. The entire section was updated and
code was corrected.

The required func cli version `v0.46.0` was added to the tutorial with a
link to the GitHub func repository tags page.

The `solution-create-bad-word-filter-service.md` was removed because it
was identical to the `create-bad-word-filter-service.md`. All references
to it were removed.

Issue #6420.

Signed-off-by: Stanislav Jakuschevskij <sjakusch@redhat.com>

* Remove func version

Signed-off-by: Stanislav Jakuschevskij <sjakusch@redhat.com>

---------

Signed-off-by: Stanislav Jakuschevskij <sjakusch@redhat.com>